+++ error.h
@@ -14,17 +14,17 @@ public:
 
     virtual void execute();
 
-    void setCSVRowNumber(int rowNumber);
-    void setCSVColumnNumber(int columnNumber);
+    void setCSVRowNumber(size_t rowNumber);
+    void setCSVColumnNumber(size_t columnNumber);
 
-    int getCSVRowNumber() const;
-    int getCSVColumnNumber() const;
+    size_t getCSVRowNumber() const;
+    size_t getCSVColumnNumber() const;
 
 protected:
     std::ofstream& mErrorStream;
     ErrorCode mErrorCode;
-    int mCSVRowNumber;              // row from parsed CSV file where the error is located
-    int mCSVColumnNumber;           // column from parsed CSV file where the error is located
+    size_t mCSVRowNumber;              // row from parsed CSV file where the error is located
+    size_t mCSVColumnNumber;           // column from parsed CSV file where the error is located
 };
 
 using ErrorPtr = std::shared_ptr<Error>;
